feat: basic functions

+++ controllers/bookController.js
@@ -1,11 +1,12 @@
 import routes from "../routes";
 import Book from "../models/Book";
-
+import User from "../models/User";
 // Home
 
 export const home = async (req, res) => {
   try {
     const books = await Book.find({}).sort({ _id: -1 });
+    console.log(books);
     res.render("home", { pageTitle: "Home", books });
   } catch (error) {
     console.log(error);
@@ -23,6 +24,7 @@ export const search = async(req, res) => {
   try {
     books = await Book.find({
       title: { $regex: searchingBy, $options: "i" }
+
     });
   } catch (error) {
     console.log(error);
@@ -36,14 +38,22 @@ export const getUpload = (req, res) =>
   res.render("upload", { pageTitle: "Upload" });
 
 export const postUpload = async (req, res) => {
+  console.log(req);
   const {
-    body: { title, description },
+    body: { title, description, author, published_date },
     file: { path }
   } = req;
   const newBook = await Book.create({
-    coverUrl: path,
+    thumbnail: path,
     title,
-    description
+    description,
+    author,
+    published_date,
+    updater: req.user._id
+  });
+  req.user.books.push(newBook.id);
+  await User.findByIdAndUpdate(req.user._id, {
+    books: req.user.books
   });
   res.redirect(routes.bookDetail(newBook.id));
 };
@@ -55,7 +65,8 @@ export const bookDetail = async (req, res) => {
     params: { id }
   } = req;
   try {
-    const book = await Book.findById(id);
+    const book =  await Book.findById(id).populate("updater");
+    console.log(book);
     res.render("bookDetail", { pageTitle: book.title, book });
   } catch (error) {
     res.redirect(routes.home);
@@ -68,9 +79,15 @@ export const getEditBook = async (req, res) => {
   const {
     params: { id }
   } = req;
+  console.log(req.user._id);
+  
   try {
     const book = await Book.findById(id);
-    res.render("editBook", { pageTitle: `Edit ${book.title}`, book });
+    if (book.updater != req.user._id) {
+      throw Error();
+    } else {
+      res.render("editBook", { pageTitle: `Book ${book.title}`, book });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -97,7 +114,12 @@ export const deleteBook = async (req, res) => {
     params: { id }
   } = req;
   try {
-    await Book.findOneAndRemove({ _id: id });
+    const book = await Book.findById(id);
+    if (book.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Book.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
